Added widget to style image upload

+++ products/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+# Check access with login decorator
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -86,8 +88,14 @@ def product_detail(request, product_id):
 
 
 # Add Product form view
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    # Limit user access 
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -110,8 +118,14 @@ def add_product(request):
 
 
 # Edit Product view
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    # Limit user access 
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -135,8 +149,14 @@ def edit_product(request, product_id):
 
 
 # Delete products
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    # Limit user access 
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
